feat(x): Phase 6 — schedulers and sync services boot in the standalone server (#905)
SEPARATION_PLAN.md Phase 6: the headless-safe half of Rowboat's boot
moves into core (boot/services.ts) — data repairs (runs migration,
code-session backfill), retention sweep, mobile channels, live-note +
bg-task schedulers, skills watcher, event processor + consumers, all
knowledge sync services (gmail/outlook/calendars/fireflies/granola),
graph builder, note tagging, inline tasks, agent runner, agent notes,
calendar notifications, meeting prep, chrome-extension sync.

Electron main and the standalone rowboat-server now call the same
prepareCoreData()/initCoreServices() in the same order — main keeps only
client-machine concerns (windows, tray, meeting detection, quick-ask,
updater, capture) plus its window fan-out watchers. The standalone
entrypoint is no longer a stub: it runs the full core, which is the
child process the Phase 7 flip will spawn.

Verified: standalone server on an isolated ROWBOAT_WORKDIR boots every
scheduler/sync with no Electron (health + RPC answering, clean
shutdown); Electron main boots identically to before on the shared
module; split-brain lock correctly refused a second host during
testing.
